add: suspense and lazy loading
