+++ project/src/components/app/app.tsx
@@ -1,10 +1,14 @@
-import MainPage from '../../pages/MainPage/MainPage';
+import MainPage from '../../pages/MainPage/ui/MainPage';
 import {RouterProvider} from 'react-router-dom';
 import {router} from '../Routers/AppRouter/AppRouter';
+import {Suspense} from 'react';
+import Loader from '../Loader/Loader';
 
 function App(): JSX.Element {
   return (
-    <RouterProvider router={router} />
+    <Suspense fallback={<Loader />} >
+      <RouterProvider router={router} />
+    </Suspense>
   );
 }
