+++ project/src/pages/AddReviewPage/index.ts
@@ -1,0 +1,1 @@
+export {AddReviewPageAsync as AddReviewPage} from './ui/AddReviewPageAsync';
