+++ project/src/pages/AddReviewPage/ui/AddReviewPageAsync.tsx
@@ -1,0 +1,3 @@
+import {lazy} from 'react';
+
+export const AddReviewPageAsync = lazy(() => import('./AddReviewPage'));
