+++ project/src/pages/FilmPage/index.ts
@@ -1,0 +1,1 @@
+export {FilmPageAsync as FilmPage} from './ui/FilmPageAsync';
