+++ project/src/pages/FilmPage/ui/FilmPageAsync.tsx
@@ -1,0 +1,3 @@
+import {lazy} from 'react';
+
+export const FilmPageAsync = lazy(() => import('./FilmPage'));
