+++ project/src/pages/MainPage/ui/MainPageAsync.tsx
@@ -1,0 +1,3 @@
+import {lazy} from 'react';
+
+export const MainPageAsync = lazy(() => import('./MainPage'));
