+++ project/src/pages/MyListPage/index.ts
@@ -1,0 +1,1 @@
+export {MyListPageAsync as MyListPage} from './ui/MyListPageAsync';
