+++ project/src/pages/MyListPage/ui/MyListPageAsync.tsx
@@ -1,0 +1,3 @@
+import {lazy} from 'react';
+
+export const MyListPageAsync = lazy(() => import('./MyListPage'));
