--- project/src/pages/NotFoundPage/NotFoundPage.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import Header from '../../components/Headers/Header';
-import Footer from '../../components/Footer/Footer';
-import GoBackNavigate from '../../components/GoBackNavigate/GoBackNavigate';
-
-const NotFoundPage = () => (
-  <div className="user-page">
-    <Header />
-    <GoBackNavigate />
-    <Footer />
-  </div>
-);
-
-export default NotFoundPage;
