+++ project/src/pages/NotFoundPage/index.ts
@@ -1,0 +1,1 @@
+export {NotFoundPageAsync as NotFoundPage} from './ui/NotFoundPageAsync';
