+++ project/src/pages/NotFoundPage/ui/NotFoundPageAsync.tsx
@@ -1,0 +1,3 @@
+import {lazy} from 'react';
+
+export const NotFoundPageAsync = lazy(() => import('./NotFoundPage'));
