+++ project/src/pages/PlayerPage/index.ts
@@ -1,0 +1,1 @@
+export {PlayerPageAsync as PlayerPage} from './ui/PlayerPageAsync';
