+++ project/src/pages/PlayerPage/ui/PlayerPage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Player from '../../components/Player/Player';
+import Player from '../../../components/Player/Player';
 
 const PlayerPage = () => (
   <Player />
