+++ project/src/pages/PlayerPage/ui/PlayerPageAsync.tsx
@@ -1,0 +1,3 @@
+import {lazy} from 'react';
+
+export const PlayerPageAsync = lazy(() => import('./PlayerPage'));
