+++ project/src/pages/SignInPage/index.ts
@@ -1,0 +1,1 @@
+export {SignInPageAsync as SignInPage} from './ui/SignInPage.Async';
