+++ project/src/pages/SignInPage/ui/SignInPage.Async.tsx
@@ -1,0 +1,3 @@
+import {lazy} from 'react';
+
+export const SignInPageAsync = lazy(() => import('./SignInPage'));
