+++ project/src/pages/SignInPage/ui/SignInPage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import Footer from '../../components/Footer/Footer';
-import SignInForm from '../../components/Forms/SignInForm/SignInForm';
-import Header from '../../components/Headers/Header';
+import Footer from '../../../components/Footer/Footer';
+import SignInForm from '../../../components/Forms/SignInForm/SignInForm';
+import Header from '../../../components/Headers/Header';
 
 const SignInPage = () => (
   <div className="user-page">
